move dagmc related methods into dagmc.py, factorise Universes method
