+++ openmc/model/model.py
@@ -1036,7 +1036,7 @@ class Model:
         """
         self.differentiate_mats(diff_volume_method, depletable_only=True)
 
-    def differentiate_depletable_mats(self, diff_volume_method: str,
+    def differentiate_mats(self, diff_volume_method: str,
                                       depletable_only: bool = True):
         """Assign distribmats for each depletable material
